+++ spec/feature/intake/higher_level_review/edit_spec.rb
@@ -1,6 +1,6 @@
 require "support/intake_helpers"
 
-feature "Edit issues" do
+feature "Edit Higher Level Review issues" do
   include IntakeHelpers
 
   before do
@@ -82,367 +82,6 @@ feature "Edit issues" do
       ]
     )
   end
-
-  def check_row(label, text)
-    row = find("tr", text: label)
-    expect(row).to have_text(text)
-  end
-
-  context "appeals" do
-    let(:legacy_opt_in_approved) { false }
-    let!(:appeal) do
-      create(:appeal,
-             veteran_file_number: veteran.file_number,
-             receipt_date: receipt_date,
-             docket_type: "evidence_submission",
-             veteran_is_not_claimant: false,
-             legacy_opt_in_approved: legacy_opt_in_approved).tap(&:create_tasks_on_intake_success!)
-    end
-
-    let(:nonrating_request_issue_attributes) do
-      {
-        review_request: appeal,
-        issue_category: "Military Retired Pay",
-        description: "nonrating description",
-        contention_reference_id: "1234",
-        decision_date: 1.month.ago
-      }
-    end
-
-    let!(:nonrating_request_issue) { create(:request_issue, nonrating_request_issue_attributes) }
-
-    let(:rating_request_issue_attributes) do
-      {
-        review_request: appeal,
-        contested_rating_issue_reference_id: "def456",
-        contested_rating_issue_profile_date: profile_date,
-        description: "PTSD denied",
-        contention_reference_id: "4567"
-      }
-    end
-
-    let!(:rating_request_issue) { create(:request_issue, rating_request_issue_attributes) }
-
-    scenario "allows adding/removing issues" do
-      visit "appeals/#{appeal.uuid}/edit/"
-
-      expect(page).to have_content(nonrating_request_issue.description)
-
-      # remove an issue
-      nonrating_intake_num = find_intake_issue_number_by_text(nonrating_request_issue.issue_category)
-      click_remove_intake_issue(nonrating_intake_num)
-      click_remove_issue_confirmation
-      expect(page).not_to have_content(nonrating_request_issue.description)
-
-      # add a different issue
-      click_intake_add_issue
-      add_intake_rating_issue("Left knee granted")
-
-      # save
-      expect(page).to have_content("Left knee granted")
-      safe_click("#button-submit-update")
-
-      # should redirect to queue
-      expect(page).to have_current_path("/queue/appeals/#{appeal.uuid}")
-
-      # going back to edit page should show those issues
-      visit "appeals/#{appeal.uuid}/edit/"
-      expect(page).to have_content("Left knee granted")
-      expect(page).not_to have_content("nonrating description")
-
-      # canceling should redirect to queue
-      click_on "Cancel"
-      expect(page).to have_current_path("/queue/appeals/#{appeal.uuid}")
-    end
-
-    scenario "allows removing and re-adding same issue" do
-      issue_description = rating_request_issue.description
-
-      visit "appeals/#{appeal.uuid}/edit/"
-
-      expect(page).to have_content(issue_description)
-      expect(page).to have_button("Save", disabled: true)
-
-      # remove
-      issue_num = find_intake_issue_number_by_text(issue_description)
-      click_remove_intake_issue(issue_num)
-      click_remove_issue_confirmation
-      expect(page).not_to have_content(issue_description)
-
-      # re-add
-      click_intake_add_issue
-      add_intake_rating_issue(issue_description, "a new comment")
-      expect(page).to have_content(issue_description)
-      expect(page).to_not have_content(
-        Constants.INELIGIBLE_REQUEST_ISSUES.duplicate_of_rating_issue_in_active_review.gsub("{review_title}", "Appeal")
-      )
-
-      # issue note was added
-      expect(page).to have_button("Save", disabled: false)
-    end
-
-    context "with multiple request issues with same data fields" do
-      let!(:duplicate_nonrating_request_issue) { create(:request_issue, nonrating_request_issue_attributes) }
-      let!(:duplicate_rating_request_issue) { create(:request_issue, rating_request_issue_attributes) }
-
-      scenario "saves by id" do
-        visit "appeals/#{appeal.uuid}/edit/"
-        expect(page).to have_content(duplicate_nonrating_request_issue.description, count: 2)
-        expect(page).to have_content(duplicate_rating_request_issue.description, count: 2)
-
-        # add another new issue
-        click_intake_add_issue
-        click_intake_no_matching_issues
-        add_intake_nonrating_issue(
-          category: "Active Duty Adjustments",
-          description: "A description!",
-          date: "04/26/2018"
-        )
-
-        safe_click("#button-submit-update")
-        safe_click ".confirm"
-        expect(page).to have_current_path("/queue/appeals/#{appeal.uuid}")
-
-        request_issue_update = RequestIssuesUpdate.where(review: appeal).last
-        non_modified_ids = [duplicate_nonrating_request_issue.id, duplicate_rating_request_issue.id,
-                            nonrating_request_issue.id, rating_request_issue.id]
-
-        # duplicate issues should be neither added nor removed
-        expect(request_issue_update.created_issues.map(&:id)).to_not include(non_modified_ids)
-        expect(request_issue_update.removed_issues.map(&:id)).to_not include(non_modified_ids)
-      end
-    end
-
-    context "with legacy appeals" do
-      before do
-        setup_legacy_opt_in_appeals(veteran.file_number)
-      end
-
-      context "with legacy_opt_in_approved" do
-        let(:legacy_opt_in_approved) { true }
-        scenario "adding issues" do
-          visit "appeals/#{appeal.uuid}/edit/"
-
-          click_intake_add_issue
-          expect(page).to have_content("Next")
-          add_intake_rating_issue("Left knee granted")
-
-          # expect legacy opt in modal
-          expect(page).to have_content("Does issue 3 match any of these VACOLS issues?")
-
-          add_intake_rating_issue("intervertebral disc syndrome") # ineligible issue
-
-          expect(page).to have_content(
-            "Left knee granted #{Constants.INELIGIBLE_REQUEST_ISSUES.legacy_appeal_not_eligible}"
-          )
-
-          click_intake_add_issue
-          add_intake_rating_issue("Back pain")
-          add_intake_rating_issue("ankylosis of hip") # eligible issue
-
-          safe_click("#button-submit-update")
-          safe_click ".confirm"
-
-          expect(page).to have_current_path("/queue/appeals/#{appeal.uuid}")
-
-          expect(RequestIssue.find_by(
-                   description: "Left knee granted",
-                   ineligible_reason: :legacy_appeal_not_eligible,
-                   vacols_id: "vacols2",
-                   vacols_sequence_id: "1"
-                 )).to_not be_nil
-
-          ri_with_optin = RequestIssue.find_by(
-            description: "Back pain",
-            ineligible_reason: nil,
-            vacols_id: "vacols1",
-            vacols_sequence_id: "1"
-          )
-
-          expect(ri_with_optin).to_not be_nil
-          li_optin = ri_with_optin.legacy_issue_optin
-          expect(li_optin.optin_processed_at).to_not be_nil
-          expect(VACOLS::CaseIssue.find_by(isskey: "vacols1", issseq: 1).issdc).to eq(
-            LegacyIssueOptin::VACOLS_DISPOSITION_CODE
-          )
-
-          # Check rollback
-          visit "appeals/#{appeal.uuid}/edit/"
-          click_remove_intake_issue_by_text("Back pain")
-          click_remove_issue_confirmation
-          safe_click("#button-submit-update")
-          safe_click ".confirm"
-
-          expect(page).to have_current_path("/queue/appeals/#{appeal.uuid}")
-          expect(li_optin.reload.rollback_processed_at).to_not be_nil
-          expect(VACOLS::CaseIssue.find_by(isskey: "vacols1", issseq: 1).issdc).to eq(
-            li_optin.original_disposition_code
-          )
-        end
-      end
-
-      context "with legacy opt in not approved" do
-        let(:legacy_opt_in_approved) { false }
-        scenario "adding issues" do
-          visit "appeals/#{appeal.uuid}/edit/"
-          click_intake_add_issue
-          add_intake_rating_issue("Left knee granted")
-
-          expect(page).to have_content("Does issue 3 match any of these VACOLS issues?")
-          # do not show inactive appeals when legacy opt in is false
-          expect(page).to_not have_content("impairment of hip")
-          expect(page).to_not have_content("typhoid arthritis")
-
-          add_intake_rating_issue("ankylosis of hip")
-
-          expect(page).to have_content(
-            "Left knee granted #{Constants.INELIGIBLE_REQUEST_ISSUES.legacy_issue_not_withdrawn}"
-          )
-
-          safe_click("#button-submit-update")
-          safe_click ".confirm"
-
-          expect(page).to have_current_path("/queue/appeals/#{appeal.uuid}")
-
-          expect(RequestIssue.find_by(
-                   description: "Left knee granted",
-                   ineligible_reason: :legacy_issue_not_withdrawn,
-                   vacols_id: "vacols1",
-                   vacols_sequence_id: "1"
-                 )).to_not be_nil
-        end
-      end
-
-      scenario "adding issue with legacy opt in disabled" do
-        allow(FeatureToggle).to receive(:enabled?).and_call_original
-        allow(FeatureToggle).to receive(:enabled?).with(:intake_legacy_opt_in, user: current_user).and_return(false)
-
-        visit "appeals/#{appeal.uuid}/edit/"
-
-        click_intake_add_issue
-        expect(page).to have_content("Add this issue")
-        add_intake_rating_issue("Left knee granted")
-        expect(page).to have_content("Left knee granted")
-      end
-    end
-  end
-
-  # rubocop:disable Metrics/MethodLength
-  # rubocop:disable Metrics/AbcSize
-  def verify_decision_issues_can_be_added_and_removed(page_url,
-                                                      original_request_issue,
-                                                      review_request,
-                                                      contested_decision_issues)
-    visit page_url
-    expect(page).to have_content("currently contesting decision issue")
-    expect(page).to have_content("PTSD denied")
-
-    # check that we cannot add the same issue again
-    click_intake_add_issue
-    expect(page).to have_css("input[disabled]", visible: false)
-    expect(page).to have_content("PTSD denied (already selected for")
-
-    nonrating_decision_issue_description = "nonrating decision issue dispositon: " \
-                                           "Active Duty Adjustments - Test nonrating decision issue"
-    rating_decision_issue_description = "a rating decision issue"
-    # check that nonrating and rating decision issues show up
-
-    expect(page).to have_content(nonrating_decision_issue_description)
-    expect(page).to have_content(rating_decision_issue_description)
-    safe_click ".close-modal"
-
-    # remove original decision issue
-    click_remove_intake_issue_by_text("currently contesting decision issue")
-    click_remove_issue_confirmation
-
-    # add new decision issue
-
-    click_intake_add_issue
-    add_intake_rating_issue(rating_decision_issue_description)
-    expect(page).to have_content(rating_decision_issue_description)
-
-    click_intake_add_issue
-
-    add_intake_rating_issue(nonrating_decision_issue_description)
-
-    binding.pry
-    expect(page).to have_content(nonrating_decision_issue_description)
-    expect(page).to have_content(
-      Constants.INELIGIBLE_REQUEST_ISSUES
-        .duplicate_of_rating_issue_in_active_review.gsub("{review_title}", "Higher-Level Review")
-    )
-
-    safe_click("#button-submit-update")
-    safe_click ".confirm"
-    expect(page).to have_current_path("/#{page_url}/confirmation")
-
-    visit page_url
-    expect(page).to have_content(nonrating_decision_issue_description)
-    expect(page).to have_content(rating_decision_issue_description)
-    expect(page).to have_content("PTSD denied")
-
-    # check that decision_request_issue is closed
-    updated_request_issue = RequestIssue.find_by(id: original_request_issue.id)
-    expect(updated_request_issue.review_request).to be_nil
-
-    # check that new request issue is created contesting the decision issue
-    request_issues = review_request.reload.request_issues
-    ri_one = request_issues.find { |ri| ri.contested_decision_issue_id == contested_decision_issues.first.id }
-    ri_two = request_issues.find { |ri| ri.contested_decision_issue_id == contested_decision_issues.second.id }
-
-    expect(ri_one.description).to eq(contested_decision_issues.first.formatted_description)
-    expect(ri_two.ineligible_reason).to eq(:duplicate_of_rating_issue_in_active_review)
-    expect(ro_two.description).to eq(contested_decision_issues.second.formatted_description)
-  end
-  # rubocop:enable Metrics/MethodLength
-  # rubocop:enable Metrics/AbcSize
-
-  # rubocop:disable Metrics/MethodLength
-  # rubocop:disable Metrics/AbcSize
-  def verify_request_issue_contending_decision_issue_not_readded(
-      page_url,
-      decision_review,
-      contested_decision_issues
-    )
-    # verify that not modifying a request issue contenting a decision issue
-    # does not result in readding
-
-    visit page_url
-    expect(page).to have_content(contested_decision_issues.first.description)
-    expect(page).to have_content(contested_decision_issues.second.description)
-    expect(page).to have_content("PTSD denied")
-
-    click_remove_intake_issue_by_text("PTSD denied")
-    click_remove_issue_confirmation
-
-    click_intake_add_issue
-    add_intake_rating_issue("Issue with legacy issue not withdrawn")
-
-    safe_click("#button-submit-update")
-    expect(page).to have_content("has been processed")
-
-    first_not_modified_request_issue = RequestIssue.find_by(
-      review_request: decision_review,
-      contested_decision_issue_id: contested_decision_issues.first.id
-    )
-
-    second_not_modified_request_issue = RequestIssue.find_by(
-      review_request: decision_review,
-      contested_decision_issue_id: contested_decision_issues.second.id
-    )
-
-    expect(first_not_modified_request_issue).to_not be_nil
-    expect(second_not_modified_request_issue).to_not be_nil
-
-    non_modified_ids = [first_not_modified_request_issue.id, second_not_modified_request_issue.id]
-    request_issue_update = RequestIssuesUpdate.find_by(review: decision_review)
-
-    # existing issues should not be added or removed
-    expect(request_issue_update.created_issues.map(&:id)).to_not include(non_modified_ids)
-    expect(request_issue_update.removed_issues.map(&:id)).to_not include(non_modified_ids)
-  end
-  # rubocop:enable Metrics/MethodLength
-  # rubocop:enable Metrics/AbcSize
 
   context "Higher-Level Reviews" do
     let!(:higher_level_review) do
@@ -1315,470 +954,6 @@ feature "Edit issues" do
           expect(page).to have_current_path("/higher_level_reviews/#{rating_ep_claim_id}/edit/cleared_eps")
           expect(page).to have_content("Issues Not Editable")
           expect(page).to have_content(Constants.INTAKE_FORM_NAMES.higher_level_review)
-        end
-      end
-    end
-  end
-
-  context "Supplemental claims" do
-    let(:is_dta_error) { false }
-    let(:benefit_type) { "compensation" }
-
-    let!(:supplemental_claim) do
-      SupplementalClaim.create!(
-        veteran_file_number: veteran.file_number,
-        receipt_date: receipt_date,
-        benefit_type: benefit_type,
-        is_dta_error: is_dta_error,
-        veteran_is_not_claimant: true
-      )
-    end
-
-    # create intake
-    let!(:intake) do
-      Intake.create!(
-        user_id: current_user.id,
-        detail: supplemental_claim,
-        veteran_file_number: veteran.file_number,
-        started_at: Time.zone.now,
-        completed_at: Time.zone.now,
-        completion_status: "success",
-        type: "SupplementalClaimIntake"
-      )
-    end
-
-    let(:rating_ep_claim_id) do
-      EndProductEstablishment.find_by(
-        source: supplemental_claim,
-        code: "040SCR"
-      ).reference_id
-    end
-
-    before do
-      supplemental_claim.create_claimants!(participant_id: "5382910292", payee_code: "10")
-
-      allow(Fakes::VBMSService).to receive(:create_contentions!).and_call_original
-
-      allow_any_instance_of(Fakes::BGSService).to receive(:find_all_relationships).and_return(
-        first_name: "BOB",
-        last_name: "VANCE",
-        ptcpnt_id: "5382910292",
-        relationship_type: "Spouse"
-      )
-    end
-
-    context "when there is a non-rating end product" do
-      let!(:nonrating_request_issue) do
-        RequestIssue.create!(
-          review_request: supplemental_claim,
-          issue_category: "Military Retired Pay",
-          description: "nonrating description",
-          contention_reference_id: "1234",
-          benefit_type: "compensation",
-          decision_date: 1.month.ago
-        )
-      end
-
-      before do
-        supplemental_claim.create_issues!([nonrating_request_issue])
-        supplemental_claim.establish!
-      end
-
-      context "when it is created due to a DTA error" do
-        let(:is_dta_error) { true }
-
-        it "cannot be edited" do
-          nonrating_dta_claim_id = EndProductEstablishment.find_by(
-            source: supplemental_claim,
-            code: "040HDENR"
-          ).reference_id
-
-          visit "supplemental_claims/#{nonrating_dta_claim_id}/edit"
-          expect(page).to have_content("Issues Not Editable")
-        end
-
-        context "when benefit type is pension" do
-          let(:benefit_type) { "pension" }
-          it "cannot be edited" do
-            nonrating_dta_claim_id = EndProductEstablishment.find_by(
-              source: supplemental_claim,
-              code: "040HDENRPMC"
-            ).reference_id
-
-            visit "supplemental_claims/#{nonrating_dta_claim_id}/edit"
-            expect(page).to have_content("Issues Not Editable")
-          end
-        end
-      end
-
-      it "shows the Supplemental Claim Edit page with a nonrating claim id" do
-        nonrating_ep_claim_id = EndProductEstablishment.find_by(
-          source: supplemental_claim,
-          code: "040SCNR"
-        ).reference_id
-        visit "supplemental_claims/#{nonrating_ep_claim_id}/edit"
-
-        expect(page).to have_content("Military Retired Pay")
-
-        click_intake_add_issue
-        click_intake_no_matching_issues
-        add_intake_nonrating_issue(
-          category: "Active Duty Adjustments",
-          description: "A description!",
-          date: "04/25/2018"
-        )
-
-        safe_click("#button-submit-update")
-
-        expect(page).to have_content("The review originally had 1 issue but now has 2.")
-        safe_click ".confirm"
-
-        expect(page).to have_current_path(
-          "/supplemental_claims/#{nonrating_ep_claim_id}/edit/confirmation"
-        )
-      end
-    end
-
-    context "when there is a rating end product" do
-      let(:request_issue) do
-        RequestIssue.create!(
-          contested_rating_issue_reference_id: "def456",
-          contested_rating_issue_profile_date: rating.profile_date,
-          review_request: supplemental_claim,
-          benefit_type: "compensation",
-          description: "PTSD denied"
-        )
-      end
-
-      let(:request_issues) { [request_issue] }
-
-      before do
-        supplemental_claim.create_issues!(request_issues)
-        supplemental_claim.establish!
-      end
-
-      context "when it is created due to a DTA error" do
-        let(:is_dta_error) { true }
-
-        it "cannot be edited" do
-          rating_dta_claim_id = EndProductEstablishment.find_by(
-            source: supplemental_claim,
-            code: "040HDER"
-          ).reference_id
-
-          visit "supplemental_claims/#{rating_dta_claim_id}/edit"
-          expect(page).to have_content("Issues Not Editable")
-        end
-
-        context "when benefit type is pension" do
-          let(:benefit_type) { "pension" }
-          it "cannot be edited" do
-            rating_dta_claim_id = EndProductEstablishment.find_by(
-              source: supplemental_claim,
-              code: "040HDERPMC"
-            ).reference_id
-
-            visit "supplemental_claims/#{rating_dta_claim_id}/edit"
-            expect(page).to have_content("Issues Not Editable")
-          end
-        end
-      end
-
-      it "shows request issues and allows adding/removing issues" do
-        visit "supplemental_claims/#{rating_ep_claim_id}/edit"
-
-        # Check that request issues appear correctly as added issues
-        expect(page).to_not have_content("Left knee granted")
-        expect(page).to have_content("PTSD denied")
-
-        expect(page).to have_content("Add / Remove Issues")
-        check_row("Form", Constants.INTAKE_FORM_NAMES.supplemental_claim)
-        check_row("Benefit type", "Compensation")
-        check_row("Claimant", "Bob Vance, Spouse (payee code 10)")
-
-        click_intake_add_issue
-
-        expect(page).to have_content("Add issue 2")
-        expect(page).to have_content("Does issue 2 match any of these issues")
-        expect(page).to have_content("Left knee granted")
-        expect(page).to have_content("PTSD denied")
-
-        # test canceling adding an issue by closing the modal
-        safe_click ".close-modal"
-        expect(page).to_not have_content("2. Left knee granted")
-
-        # adding an issue should show the issue
-        click_intake_add_issue
-        add_intake_rating_issue("Left knee granted")
-
-        expect(page).to have_content("2. Left knee granted")
-        expect(page).to_not have_content("Notes:")
-        click_remove_intake_issue("1")
-
-        # expect a pop up
-        expect(page).to have_content("Are you sure you want to remove this issue?")
-        click_remove_issue_confirmation
-
-        expect(page).not_to have_content("PTSD denied")
-
-        # re-add to proceed
-        click_intake_add_issue
-        add_intake_rating_issue("PTSD denied", "I am an issue note")
-        expect(page).to have_content("2. PTSD denied")
-        expect(page).to have_content("I am an issue note")
-
-        # clicking add issue again should show a disabled radio button for that same rating
-        click_intake_add_issue
-        expect(page).to have_content("Add issue 3")
-        expect(page).to have_content("Does issue 3 match any of these issues")
-        expect(page).to have_content("Left knee granted (already selected for issue 1)")
-        expect(page).to have_css("input[disabled]", visible: false)
-
-        # Add nonrating issue
-        click_intake_no_matching_issues
-        add_intake_nonrating_issue(
-          category: "Active Duty Adjustments",
-          description: "Description for Active Duty Adjustments",
-          date: "04/25/2018"
-        )
-        expect(page).to have_content("3 issues")
-
-        # add unidentified issue
-        click_intake_add_issue
-        add_intake_unidentified_issue("This is an unidentified issue")
-        expect(page).to have_content("4 issues")
-        expect(page).to have_content("This is an unidentified issue")
-      end
-
-      context "when veteran has active nonrating request issues" do
-        let(:another_higher_level_review) do
-          create(:higher_level_review,
-                 veteran_file_number: veteran.file_number,
-                 benefit_type: "compensation")
-        end
-
-        let!(:active_nonrating_request_issue) do
-          create(:request_issue,
-                 :nonrating,
-                 review_request: another_higher_level_review)
-        end
-
-        before do
-          another_higher_level_review.create_issues!([active_nonrating_request_issue])
-        end
-
-        scenario "shows ineligibility message and saves conflicting request issue id" do
-          visit "supplemental_claims/#{rating_ep_claim_id}/edit"
-          click_intake_add_issue
-          click_intake_no_matching_issues
-
-          fill_in "Issue category", with: active_nonrating_request_issue.issue_category
-          find("#issue-category").send_keys :enter
-          expect(page).to have_content("Does issue 2 match any of the issues actively being reviewed?")
-          expect(page).to have_content("#{active_nonrating_request_issue.issue_category}: " \
-                                       "#{active_nonrating_request_issue.description}")
-          add_active_intake_nonrating_issue(active_nonrating_request_issue.issue_category)
-          expect(page).to have_content("#{active_nonrating_request_issue.issue_category} -" \
-                                       " #{active_nonrating_request_issue.description}" \
-                                       " is ineligible because it's already under review as a Higher-Level Review")
-
-          safe_click("#button-submit-update")
-          safe_click ".confirm"
-          expect(page).to have_current_path(
-            "/supplemental_claims/#{rating_ep_claim_id}/edit/confirmation"
-          )
-
-          expect(RequestIssue.find_by(review_request: supplemental_claim,
-                                      issue_category: active_nonrating_request_issue.issue_category,
-                                      ineligible_due_to: active_nonrating_request_issue.id,
-                                      ineligible_reason: "duplicate_of_nonrating_issue_in_active_review",
-                                      description: active_nonrating_request_issue.description,
-                                      decision_date: active_nonrating_request_issue.decision_date)).to_not be_nil
-        end
-      end
-
-      context "has decision issues" do
-        let(:contested_decision_issues) { setup_prior_decision_issues(veteran) }
-        let(:decision_request_issue) do
-          create(
-            :request_issue,
-            review_request: supplemental_claim,
-            description: "currently contesting decision issue",
-            decision_date: Time.zone.now - 2.days,
-            contested_decision_issue_id: contested_decision_issues.first.id
-          )
-        end
-
-        let(:request_issues) { [request_issue, decision_request_issue] }
-
-        let!(:request_issue_that_causes_ineligiblity) do
-          already_active_hlr = create(:higher_level_review, :with_end_product_establishment)
-          create(
-            :request_issue,
-            review_request: already_active_hlr,
-            description: "currently active request issue",
-            decision_date: Time.zone.now - 2.days,
-            end_product_establishment_id: already_active_hlr.end_product_establishments.first.id,
-            contested_decision_issue_id: contested_decision_issues.second.id
-          )
-        end
-
-        it "shows decision isssues and allows adding/removing issues" do
-          verify_decision_issues_can_be_added_and_removed(
-            "supplemental_claims/#{rating_ep_claim_id}/edit",
-            decision_request_issue,
-            supplemental_claim,
-            contested_decision_issues
-          )
-        end
-      end
-
-      context "with existing request issues contesting decision issues" do
-        let(:decision_request_issue) do
-          setup_request_issue_with_nonrating_decision_issue(supplemental_claim)
-        end
-
-        let(:nonrating_decision_request_issue) do
-          setup_request_issue_with_rating_decision_issue(
-            supplemental_claim,
-            contested_rating_issue_reference_id: "abc123"
-          )
-        end
-
-        let(:request_issues) { [request_issue, decision_request_issue, nonrating_decision_request_issue] }
-
-        it "does not remove & readd unedited issues" do
-          verify_request_issue_contending_decision_issue_not_readded(
-            "supplemental_claims/#{rating_ep_claim_id}/edit",
-            supplemental_claim,
-            decision_request_issue.decision_issues + nonrating_decision_request_issue.decision_issues
-          )
-        end
-      end
-
-      it "enables save button only when dirty" do
-        visit "supplemental_claims/#{rating_ep_claim_id}/edit"
-
-        expect(page).to have_button("Save", disabled: true)
-
-        click_intake_add_issue
-        add_intake_rating_issue("Left knee granted")
-
-        expect(page).to have_button("Save", disabled: false)
-
-        click_remove_intake_issue("2")
-        click_remove_issue_confirmation
-
-        expect(page).to_not have_content("Left knee granted")
-        expect(page).to have_button("Save", disabled: true)
-      end
-
-      it "Does not allow save if no issues are selected" do
-        visit "supplemental_claims/#{rating_ep_claim_id}/edit"
-        click_remove_intake_issue("1")
-        click_remove_issue_confirmation
-
-        expect(page).to have_button("Save", disabled: true)
-      end
-
-      scenario "shows error message if an update is in progress" do
-        RequestIssuesUpdate.create!(
-          review: supplemental_claim,
-          user: current_user,
-          before_request_issue_ids: [request_issue.id],
-          after_request_issue_ids: [request_issue.id],
-          attempted_at: Time.zone.now,
-          submitted_at: Time.zone.now,
-          processed_at: nil
-        )
-
-        visit "supplemental_claims/#{rating_ep_claim_id}/edit"
-        click_intake_add_issue
-        add_intake_rating_issue("Left knee granted")
-        safe_click("#button-submit-update")
-
-        expect(page).to have_content("The review originally had 1 issue but now has 2.")
-        safe_click ".confirm"
-
-        expect(page).to have_content("Previous update not yet done processing")
-      end
-
-      it "updates selected issues" do
-        allow(Fakes::VBMSService).to receive(:establish_claim!).and_call_original
-        allow(Fakes::VBMSService).to receive(:create_contentions!).and_call_original
-        allow(Fakes::VBMSService).to receive(:associate_rating_request_issues!).and_call_original
-        allow(Fakes::VBMSService).to receive(:remove_contention!).and_call_original
-
-        visit "supplemental_claims/#{rating_ep_claim_id}/edit"
-        click_remove_intake_issue("1")
-        click_remove_issue_confirmation
-        click_intake_add_issue
-        add_intake_rating_issue("Left knee granted")
-
-        expect(page).to have_button("Save", disabled: false)
-
-        safe_click("#button-submit-update")
-
-        expect(page).to have_current_path(
-          "/supplemental_claims/#{rating_ep_claim_id}/edit/confirmation"
-        )
-
-        # reload to verify that the new issues populate the form
-        visit "supplemental_claims/#{rating_ep_claim_id}/edit"
-        expect(page).to have_content("Left knee granted")
-        expect(page).to_not have_content("PTSD denied")
-
-        # assert server has updated data
-        new_request_issue = supplemental_claim.reload.request_issues.first
-        expect(new_request_issue.description).to eq("Left knee granted")
-        expect(request_issue.reload.review_request_id).to be_nil
-        expect(request_issue.removed_at).to eq(Time.zone.now)
-        expect(new_request_issue.rating_issue_associated_at).to eq(Time.zone.now)
-
-        # expect contentions to reflect issue update
-        expect(Fakes::VBMSService).to have_received(:create_contentions!).with(
-          veteran_file_number: veteran.file_number,
-          claim_id: rating_ep_claim_id,
-          contention_descriptions: ["Left knee granted"],
-          special_issues: [],
-          user: current_user
-        )
-        expect(Fakes::VBMSService).to have_received(:associate_rating_request_issues!).with(
-          claim_id: rating_ep_claim_id,
-          rating_issue_contention_map: {
-            new_request_issue.contested_rating_issue_reference_id => new_request_issue.contention_reference_id
-          }
-        )
-        expect(Fakes::VBMSService).to have_received(:remove_contention!).once
-      end
-
-      feature "cancel edits" do
-        def click_cancel(visit_page)
-          visit "supplemental_claims/#{rating_ep_claim_id}/edit#{visit_page}"
-          click_on "Cancel"
-          correct_path = "/supplemental_claims/#{rating_ep_claim_id}/edit/cancel"
-          expect(page).to have_current_path(correct_path)
-          expect(page).to have_content("Edit Canceled")
-          expect(page).to have_content(Constants.INTAKE_FORM_NAMES.supplemental_claim)
-        end
-
-        scenario "from landing page" do
-          click_cancel("/")
-        end
-      end
-
-      feature "with cleared end product" do
-        let!(:cleared_end_product) do
-          create(:end_product_establishment,
-                 source: supplemental_claim,
-                 synced_status: "CLR")
-        end
-
-        scenario "prevents edits on eps that have cleared" do
-          visit "supplemental_claims/#{rating_ep_claim_id}/edit/"
-          expect(page).to have_current_path("/supplemental_claims/#{rating_ep_claim_id}/edit/cleared_eps")
-          expect(page).to have_content("Issues Not Editable")
-          expect(page).to have_content(Constants.INTAKE_FORM_NAMES.supplemental_claim)
         end
       end
     end
